mdao fixswap奖励调整

+++ contracts/mdao.fixswap/include/mdao.fixswapdb.hpp
@@ -58,7 +58,7 @@ namespace wasm
         set<name> supported_tokens;
 
         uint64_t farm_lease_id          = 0;
-        map <extended_symbol, uint32_t> farm_scales;
+        map <extended_symbol, int64_t> farm_scales;
 
         EOSLIB_SERIALIZE(gswap_t, (status)(fee_collector)(admin)(make_fee_ratio)(take_fee_ratio)(swap_id)(supported_tokens)(farm_lease_id)(farm_scales))
     };
